Modify gravity mechanic, clean code, modify jump
Gravity now works as an acceleration instead of a set value like it did previously.
Jumping now works similar Worms, jump gives a set horizontal speed instead of being able to control the character midjump.
Moved several Point declarations into Getters that will return the correct value depending on whether the character is inverted.  This removes the necessity for a lot of the if (inverted) calls.
Change vectors (gravity, ground, movespeed, etc) to be time based instead of a set number of pixels.`